chore: Add .gitignore file
This change adds a .gitignore file to the repository.

The .gitignore file is configured to ignore files and directories
generated by JetBrains IDEs (PyCharm, IntelliJ, etc.), Gradle, Maven,
and other build tools. This helps to keep the repository clean and
prevents unnecessary files from being committed. It also includes an
exclusion patch that allows for code style settings and run
configurations to be tracked within the .idea directory.